+++ src/verb_invocation.rs
@@ -2,23 +2,23 @@ use regex::Regex;
 
 #[derive(Clone, Debug)]
 pub struct VerbInvocation {
-    pub key: String, // this "key" name starts to be confusing... any idea?
+    pub name: String,
     pub args: Option<String>,
 }
 impl VerbInvocation {
     pub fn from(invocation: &str) -> VerbInvocation {
         let caps = regex!(r"^(\S*)\s*(.+?)?\s*$").captures(invocation).unwrap();
-        let key = caps.get(1).unwrap().as_str().to_string();
+        let name = caps.get(1).unwrap().as_str().to_string();
         let args = caps.get(2).map(|c| c.as_str().to_string());
-        VerbInvocation { key, args }
+        VerbInvocation { name, args }
     }
     pub fn is_empty(&self) -> bool {
-        self.key.is_empty()
+        self.name.is_empty()
     }
-    pub fn to_string_for_key(&self, key: &str) -> String {
+    pub fn to_string_for_name(&self, name: &str) -> String {
         match self.args {
-            Some(ref args) => format!("{} {}", key, args),
-            None => key.to_owned(),
+            Some(ref args) => format!("{} {}", name, args),
+            None => name.to_owned(),
         }
     }
 }