+++ src/verb_store.rs
@@ -277,8 +277,8 @@ impl VerbStore {
                     continue;
                 }
             }
-            if verb.invocation.key.starts_with(prefix) {
-                if verb.invocation.key == prefix {
+            if verb.invocation.name.starts_with(prefix) {
+                if verb.invocation.name == prefix {
                     return PrefixSearchResult::Match(&verb);
                 }
                 found_index = index;
@@ -293,17 +293,17 @@ impl VerbStore {
     }
     /// return the index of the verb having the long name. This function is meant
     /// for internal access when it's sure it can't failed (i.e. for a builtin)
-    /// It looks for verbs by key, starting from the builtins, to
+    /// It looks for verbs by name, starting from the builtins, to
     /// ensure it hasn't been overriden.
     pub fn index_of(&self, name: &str) -> usize {
         for i in 0..self.verbs.len() {
-            if self.verbs[i].invocation.key == name {
+            if self.verbs[i].invocation.name == name {
                 return i;
             }
         }
         panic!("invalid verb search");
     }
-    /// return the index of the verb which is triggered by the given key, if any
+    /// return the index of the verb which is triggered by the given keyboard key, if any
     pub fn index_of_key(&self, key: KeyEvent) -> Option<usize> {
         for i in 0..self.verbs.len() {
             if let Some(verb_key) = self.verbs[i].key {